requirements.txt
piper-tts
wave
openai
moviepy